MEMC model configurations for MEND Wang et al. 2013

MEMC model configurations for MEND Wang et al. 2013

COMISSION_model

Format

an object created form configure_model.

name

COMISSION

params

the default MEMC parameter table

state

the default inital state values for MEMC

carbon_pools_func

carbon pools based on Wang et al. 2013

carbon_fluxes_func

TBD

References

Wang et al. 2013

update with the comission documetation

See also

Examples

COMISSION_model
#> $name
#> [1] "COMISSION"
#> 
#> $config
#> 
#> 
#> |Model     |POMdecomp |
#> |:---------|:---------|
#> |COMISSION |RMM       |
#> 
#> $params
#>    parameter                                             description
#> 1        V.p         maximum specific decomposition rate for P by EP
#> 2        K.p         half-saturation constant for decomposition of P
#> 3        V.m         maximum specific decomposition rate for M by EM
#> 4        K.m         half-saturation constant for decomposition of M
#> 5        V.d       maximum specific uptake rate of D for growth of B
#> 6        K.d half-saturation constant of uptake of D for growth of B
#> 7        m.r                     specific maintenance factor or rate
#> 8        E.c                                   carbon use efficiency
#> 9        f.d                 fraction of decomposed P allocated to D
#> 10       g.d                       fraction of dead B allocated to D
#> 11      p.ep                     fraction of mR for production of EP
#> 12      p.em                     fraction of mR for production of EM
#> 13      r.ep                                     turnover rate of EP
#> 14      r.em                                     turnover rate of EM
#> 15     Q.max                           maximum DOC sorption capacity
#> 16     K.ads                                specific adsorption rate
#> 17     K.des                                         desorption rate
#> 18      K.ba                                        binding affinity
#> 19       I.p                                         input rate of P
#> 20       I.d                                         input rate of D
#> 21      fI.d                                       ratio of ID to IP
#>              units   value
#> 1  mgC mgC^-1 h^-1 2.5e+00
#> 2     mgC / g soil 5.0e+01
#> 3  mgC mgC^-1 h^-1 1.0e+00
#> 4      mg C/g soil 2.5e+02
#> 5  mgC mgC^-1 h^-1 5.0e-04
#> 6      mg C/g soil 2.6e-01
#> 7  mgC mgC^-1 h^-1 2.8e-04
#> 8             <NA> 4.7e-01
#> 9             <NA> 5.0e-01
#> 10            <NA> 5.0e-01
#> 11            <NA> 1.0e-02
#> 12            <NA> 1.0e-02
#> 13 mgC mgC^-1 h^-1 1.0e-03
#> 14 mgC mgC^-1 h^-1 1.0e-03
#> 15    mgC / g soil 1.7e+00
#> 16 mgC mgC^-1 h^-1 6.0e-03
#> 17 mgC mgC^-1 h^-1 1.0e-03
#> 18   (mgC/soil)^-1 6.0e+00
#> 19 mgC mgC^-1 h^-1 8.0e-05
#> 20 mgC mgC^-1 h^-1 8.0e-05
#> 21            <NA> 1.0e+00
#> 
#> $state
#>        P        M        Q        B        D       EP       EM       IC 
#> 10.00000  5.00000  0.10000  2.00000  1.00000  0.00001  0.00001  0.00000 
#>      Tot 
#> 18.10002 
#> 
#> $env
#> <environment: 0x7f81017888f8>
#> 
#> $carbon_pools_func
#> function(t, env, state = NULL, params = NULL, flux_function = carbon_fluxes(POMdecomp = "MM")){
#> 
#>   assert_that(is.numeric(t))
#>   fx <- flux_function
#>   assert_that(is.list(flux_function))
#>   assert_that(length(fx) == 2)
#>   assert_that(is.function(fx[["flux_function"]]))
#> 
#>   # Update model parameter & initial state values if needed.
#>   if(!is.null(params)){
#>     p <- params$value
#>     names(p) <- params$parameter
#>     modify_env(env, p)
#> 
#>   }
#>   if(!is.null(state)){
#>     modify_env(env, state)
#>   }
#> 
#>   with(as.list(env),{
#> 
#>     # Define the fluxes and check to make sure they meet the requirements to be used
#>     # by the MEND carbon pool structure.
#>     #ff <- flux_function(env)[["flux_function"]]
#>     #fluxes <- ff(env)
#>     ff <- fx[["flux_function"]]
#>     fluxes <- ff(env)
#> 
#>     expected_fluxes <- rep('F', length.out = length(1:8))
#>     expected_fluxes <- c(paste0(expected_fluxes, 1:8), 'F9.ep', 'F9.em', 'F10.ep', 'F10.em')
#>     assertthat::assert_that(assertthat::has_name(x = fluxes, which = expected_fluxes))
#>     assertthat::assert_that(all(unlist(lapply(fluxes, is.function))), msg = 'fluxes input must be a list of functions')
#> 
#>     # Define the system of differential equations to describes the
#>     # changes in the carbon pool states.
#>     # -----------------------------------------------------------
#>     # P = particulate organic carbon
#>     dP <- I.p + (1 - g.d) * fluxes$F8() - fluxes$F2()
#>     # M = mineral-associated organic carbon (MOC)
#>     dM <- (1 - f.d) * fluxes$F2() - fluxes$F3()
#>     # Q = active layer of MOC
#>     dQ <- fluxes$F6() - fluxes$F7()
#>     # B = microbial biomass carbon
#>     dB <- fluxes$F1() - (fluxes$F4() + fluxes$F5()) - fluxes$F8() - (fluxes$F9.ep() + fluxes$F9.em())
#>     # D = dissolved organic carbon
#>     dD <- I.d + f.d * fluxes$F2() + g.d * fluxes$F8() + fluxes$F3() + (fluxes$F10.em() + fluxes$F10.ep()) - fluxes$F1() - (fluxes$F6() - fluxes$F7())
#>     # EP = carbon stored as extra-cellular enzymes
#>     dEP <- fluxes$F9.em() - fluxes$F10.ep()
#>     # EM = carbon stored as extra-cellular enzymes
#>     dEM <- fluxes$F9.em() - fluxes$F10.em()
#>     # IC = inorganic carbon (CO2)
#>     dIC <- fluxes$F4() + fluxes$F5()
#>     # Tot = the total carbon pool
#>     dTot <- I.p + I.d - (fluxes$F4() + fluxes$F5())
#> 
#>     # Return outputs
#>     list(c(dP, dM, dQ, dB, dD, dEP, dEM, dIC, dTot))
#>   })
#> }
#> <environment: namespace:MEMC>
#> 
#> $carbon_fluxes_func
#> $carbon_fluxes_func$flux_table
#>   POMdecomp
#> 1       RMM
#> 
#> $carbon_fluxes_func$flux_function
#> function(env, state=NULL, params=NULL){
#> 
#>     # Update model parameter & initial state values if needed.
#>     if(!is.null(params)){
#>       p <- params$value
#>       names(p) <- params$parameter
#>       modify_env(env, p)
#> 
#>     }
#>     if(!is.null(state)){
#>       modify_env(env, state)
#>     }
#>     with(as.list(env), {
#> 
#>       # Set up the original functions.
#>       fluxes  <- carbon_fluxes_internal(env = env)
#> 
#>       if(POMdecomp == "MM"){
#>         fluxes[["F1"]] = function(){
#>           # DOC uptake by microbial biomass.
#>           # Michaelis menten kinetics
#>           (1/E.c) * (V.d + m.r) * B * D /(K.d + D)
#>         }
#>       } else if(POMdecomp == "RMM") {
#>         fluxes[["F1"]] = function(){
#>           # DOC uptake by microbial biomass.
#>           # Reverse michaelis menten kinetics
#>           (1/E.c) * (V.d + m.r) * B * D /(K.d + B)
#>         }
#>       } else if(POMdecomp == "ECA"){
#>         fluxes[["F1"]] = function(){
#>           # DOC uptake by microbial biomass, note that this uses ECA kinetics.
#>           (1/E.c) * (V.d + m.r) * B * D /(K.d + B + D)
#>         }
#>       } else if(POMdecomp == "LM") {
#>         message("needs to be implemented")
#>       }
#> 
#>       return(fluxes)
#> 
#>     })
#>   }
#> <bytecode: 0x7f8106838288>
#> <environment: 0x7f81017e4f10>
#> 
#> 
solve_model(mod = COMISSION_model, time = 1:10)
#>     time variable        value       units      name
#>  1:    1        P 1.000000e+01 mg C/g soil COMISSION
#>  2:    2        P 1.000035e+01 mg C/g soil COMISSION
#>  3:    3        P 1.000070e+01 mg C/g soil COMISSION
#>  4:    4        P 1.000104e+01 mg C/g soil COMISSION
#>  5:    5        P 1.000138e+01 mg C/g soil COMISSION
#>  6:    6        P 1.000172e+01 mg C/g soil COMISSION
#>  7:    7        P 1.000206e+01 mg C/g soil COMISSION
#>  8:    8        P 1.000239e+01 mg C/g soil COMISSION
#>  9:    9        P 1.000273e+01 mg C/g soil COMISSION
#> 10:   10        P 1.000306e+01 mg C/g soil COMISSION
#> 11:    1        M 5.000000e+00 mg C/g soil COMISSION
#> 12:    2        M 5.000002e+00 mg C/g soil COMISSION
#> 13:    3        M 5.000006e+00 mg C/g soil COMISSION
#> 14:    4        M 5.000010e+00 mg C/g soil COMISSION
#> 15:    5        M 5.000016e+00 mg C/g soil COMISSION
#> 16:    6        M 5.000023e+00 mg C/g soil COMISSION
#> 17:    7        M 5.000030e+00 mg C/g soil COMISSION
#> 18:    8        M 5.000039e+00 mg C/g soil COMISSION
#> 19:    9        M 5.000049e+00 mg C/g soil COMISSION
#> 20:   10        M 5.000060e+00 mg C/g soil COMISSION
#> 21:    1        Q 1.000000e-01 mg C/g soil COMISSION
#> 22:    2        Q 1.055580e-01 mg C/g soil COMISSION
#> 23:    3        Q 1.110561e-01 mg C/g soil COMISSION
#> 24:    4        Q 1.164952e-01 mg C/g soil COMISSION
#> 25:    5        Q 1.218760e-01 mg C/g soil COMISSION
#> 26:    6        Q 1.271995e-01 mg C/g soil COMISSION
#> 27:    7        Q 1.324666e-01 mg C/g soil COMISSION
#> 28:    8        Q 1.376780e-01 mg C/g soil COMISSION
#> 29:    9        Q 1.428346e-01 mg C/g soil COMISSION
#> 30:   10        Q 1.479372e-01 mg C/g soil COMISSION
#> 31:    1        B 2.000000e+00 mg C/g soil COMISSION
#> 32:    2        B 1.999509e+00 mg C/g soil COMISSION
#> 33:    3        B 1.999010e+00 mg C/g soil COMISSION
#> 34:    4        B 1.998504e+00 mg C/g soil COMISSION
#> 35:    5        B 1.997991e+00 mg C/g soil COMISSION
#> 36:    6        B 1.997471e+00 mg C/g soil COMISSION
#> 37:    7        B 1.996944e+00 mg C/g soil COMISSION
#> 38:    8        B 1.996411e+00 mg C/g soil COMISSION
#> 39:    9        B 1.995870e+00 mg C/g soil COMISSION
#> 40:   10        B 1.995323e+00 mg C/g soil COMISSION
#> 41:    1        D 1.000000e+00 mg C/g soil COMISSION
#> 42:    2        D 9.933355e-01 mg C/g soil COMISSION
#> 43:    3        D 9.867420e-01 mg C/g soil COMISSION
#> 44:    4        D 9.802184e-01 mg C/g soil COMISSION
#> 45:    5        D 9.737639e-01 mg C/g soil COMISSION
#> 46:    6        D 9.673773e-01 mg C/g soil COMISSION
#> 47:    7        D 9.610577e-01 mg C/g soil COMISSION
#> 48:    8        D 9.548043e-01 mg C/g soil COMISSION
#> 49:    9        D 9.486160e-01 mg C/g soil COMISSION
#> 50:   10        D 9.424921e-01 mg C/g soil COMISSION
#> 51:    1       EP 1.000000e-05 mg C/g soil COMISSION
#> 52:    2       EP 1.558652e-05 mg C/g soil COMISSION
#> 53:    3       EP 2.116607e-05 mg C/g soil COMISSION
#> 54:    4       EP 2.673864e-05 mg C/g soil COMISSION
#> 55:    5       EP 3.230421e-05 mg C/g soil COMISSION
#> 56:    6       EP 3.786278e-05 mg C/g soil COMISSION
#> 57:    7       EP 4.341432e-05 mg C/g soil COMISSION
#> 58:    8       EP 4.895883e-05 mg C/g soil COMISSION
#> 59:    9       EP 5.449630e-05 mg C/g soil COMISSION
#> 60:   10       EP 6.002671e-05 mg C/g soil COMISSION
#> 61:    1       EM 1.000000e-05 mg C/g soil COMISSION
#> 62:    2       EM 1.558652e-05 mg C/g soil COMISSION
#> 63:    3       EM 2.116607e-05 mg C/g soil COMISSION
#> 64:    4       EM 2.673864e-05 mg C/g soil COMISSION
#> 65:    5       EM 3.230421e-05 mg C/g soil COMISSION
#> 66:    6       EM 3.786278e-05 mg C/g soil COMISSION
#> 67:    7       EM 4.341432e-05 mg C/g soil COMISSION
#> 68:    8       EM 4.895883e-05 mg C/g soil COMISSION
#> 69:    9       EM 5.449630e-05 mg C/g soil COMISSION
#> 70:   10       EM 6.002671e-05 mg C/g soil COMISSION
#> 71:    1       IC 0.000000e+00 mg C/g soil COMISSION
#> 72:    2       IC 1.395014e-03 mg C/g soil COMISSION
#> 73:    3       IC 2.787753e-03 mg C/g soil COMISSION
#> 74:    4       IC 4.178213e-03 mg C/g soil COMISSION
#> 75:    5       IC 5.566392e-03 mg C/g soil COMISSION
#> 76:    6       IC 6.952285e-03 mg C/g soil COMISSION
#> 77:    7       IC 8.335888e-03 mg C/g soil COMISSION
#> 78:    8       IC 9.717199e-03 mg C/g soil COMISSION
#> 79:    9       IC 1.109621e-02 mg C/g soil COMISSION
#> 80:   10       IC 1.247293e-02 mg C/g soil COMISSION
#> 81:    1      Tot 1.810002e+01 mg C/g soil COMISSION
#> 82:    2      Tot 1.809878e+01 mg C/g soil COMISSION
#> 83:    3      Tot 1.809755e+01 mg C/g soil COMISSION
#> 84:    4      Tot 1.809632e+01 mg C/g soil COMISSION
#> 85:    5      Tot 1.809509e+01 mg C/g soil COMISSION
#> 86:    6      Tot 1.809387e+01 mg C/g soil COMISSION
#> 87:    7      Tot 1.809264e+01 mg C/g soil COMISSION
#> 88:    8      Tot 1.809142e+01 mg C/g soil COMISSION
#> 89:    9      Tot 1.809020e+01 mg C/g soil COMISSION
#> 90:   10      Tot 1.808899e+01 mg C/g soil COMISSION
#>     time variable        value       units      name