MEMC model configurations

MEMC model configurations

Usage

CORPSE_model

Format

an object created form configure_model.

name

CORPSE, the model name

table

table describing the user selected kinetics

params

the default MEMC parameter table

state

the default inital state values for MEMC

See also

Examples

CORPSE_model
#> $name
#> [1] "CORPSE"
#> 
#> $table
#>    model DOMdecomp POMdecomp MBdecay
#> 1 CORPSE       RMM        LM      DD
#> 
#> $params
#>    parameter                                             description
#> 1        V.p         maximum specific decomposition rate for P by EP
#> 2        K.p         half-saturation constant for decomposition of P
#> 3        V.m         maximum specific decomposition rate for M by EM
#> 4        K.m         half-saturation constant for decomposition of M
#> 5        V.d       maximum specific uptake rate of D for growth of B
#> 6        K.d half-saturation constant of uptake of D for growth of B
#> 7        f.d                 fraction of decomposed P allocated to D
#> 8        g.d                       fraction of dead B allocated to D
#> 9       p.ep                     fraction of mR for production of EP
#> 10      p.em                     fraction of mR for production of EM
#> 11      r.ep                                     turnover rate of EP
#> 12      r.em                                     turnover rate of EM
#> 13     Q.max                           maximum DOC sorption capacity
#> 14     K.ads                                specific adsorption rate
#> 15     K.des                                         desorption rate
#> 16   dd.beta           strength of density dependent microbial decay
#> 17   Input.P                                              POM input 
#> 18   Input.D                                               DOM input
#> 19   Input.M                                                M input 
#> 20       CUE                                   carbon use efficiency
#>              units   value
#> 1  mgC mgC^-1 h^-1  14.000
#> 2     mgC / g soil  50.000
#> 3  mgC mgC^-1 h^-1   0.250
#> 4      mg C/g soil 250.000
#> 5  mgC mgC^-1 h^-1   3.000
#> 6      mg C/g soil   0.250
#> 7             <NA>   0.500
#> 8             <NA>   0.500
#> 9             <NA>   0.010
#> 10            <NA>   0.010
#> 11 mgC mgC^-1 h^-1   0.001
#> 12 mgC mgC^-1 h^-1   0.001
#> 13    mgC / g soil   3.400
#> 14 mgC mgC^-1 h^-1   0.006
#> 15 mgC mgC^-1 h^-1   0.001
#> 16            <NA>   2.000
#> 17            mg C   0.000
#> 18            mg C   0.000
#> 19            mg C   0.000
#> 20                   0.400
#> 
#> $state
#>        P        M        Q        B        D       EP       EM       IC 
#> 10.00000  5.00000  0.10000  2.00000  1.00000  0.00001  0.00001  0.00000 
#>      Tot 
#> 18.10002 
#> 
solve_model(mod = CORPSE_model, time = 1:10)
#>     time variable        value       units   name
#>  1:    1        P 1.000000e+01 mg C/g soil CORPSE
#>  2:    2        P 8.171658e-02 mg C/g soil CORPSE
#>  3:    3        P 2.815842e-02 mg C/g soil CORPSE
#>  4:    4        P 1.233516e-02 mg C/g soil CORPSE
#>  5:    5        P 6.471172e-03 mg C/g soil CORPSE
#>  6:    6        P 3.847193e-03 mg C/g soil CORPSE
#>  7:    7        P 2.497051e-03 mg C/g soil CORPSE
#>  8:    8        P 1.727448e-03 mg C/g soil CORPSE
#>  9:    9        P 1.253683e-03 mg C/g soil CORPSE
#> 10:   10        P 9.435658e-04 mg C/g soil CORPSE
#> 11:    1        M 5.000000e+00 mg C/g soil CORPSE
#> 12:    2        M 1.084021e+01 mg C/g soil CORPSE
#> 13:    3        M 1.118479e+01 mg C/g soil CORPSE
#> 14:    4        M 1.131623e+01 mg C/g soil CORPSE
#> 15:    5        M 1.137861e+01 mg C/g soil CORPSE
#> 16:    6        M 1.141310e+01 mg C/g soil CORPSE
#> 17:    7        M 1.143427e+01 mg C/g soil CORPSE
#> 18:    8        M 1.144823e+01 mg C/g soil CORPSE
#> 19:    9        M 1.145791e+01 mg C/g soil CORPSE
#> 20:   10        M 1.146488e+01 mg C/g soil CORPSE
#> 21:    1        Q 1.000000e-01 mg C/g soil CORPSE
#> 22:    2        Q 1.160005e-01 mg C/g soil CORPSE
#> 23:    3        Q 1.204555e-01 mg C/g soil CORPSE
#> 24:    4        Q 1.220711e-01 mg C/g soil CORPSE
#> 25:    5        Q 1.228354e-01 mg C/g soil CORPSE
#> 26:    6        Q 1.232681e-01 mg C/g soil CORPSE
#> 27:    7        Q 1.235434e-01 mg C/g soil CORPSE
#> 28:    8        Q 1.237324e-01 mg C/g soil CORPSE
#> 29:    9        Q 1.238691e-01 mg C/g soil CORPSE
#> 30:   10        Q 1.239715e-01 mg C/g soil CORPSE
#> 31:    1        B 2.000000e+00 mg C/g soil CORPSE
#> 32:    2        B 1.335043e+00 mg C/g soil CORPSE
#> 33:    3        B 7.908910e-01 mg C/g soil CORPSE
#> 34:    4        B 5.277281e-01 mg C/g soil CORPSE
#> 35:    5        B 3.843792e-01 mg C/g soil CORPSE
#> 36:    6        B 2.975153e-01 mg C/g soil CORPSE
#> 37:    7        B 2.403520e-01 mg C/g soil CORPSE
#> 38:    8        B 2.003064e-01 mg C/g soil CORPSE
#> 39:    9        B 1.708797e-01 mg C/g soil CORPSE
#> 40:   10        B 1.484372e-01 mg C/g soil CORPSE
#> 41:    1        D 1.000000e+00 mg C/g soil CORPSE
#> 42:    2        D 1.334209e+00 mg C/g soil CORPSE
#> 43:    3        D 4.289360e-01 mg C/g soil CORPSE
#> 44:    4        D 1.873911e-01 mg C/g soil CORPSE
#> 45:    5        D 1.020991e-01 mg C/g soil CORPSE
#> 46:    6        D 6.455758e-02 mg C/g soil CORPSE
#> 47:    7        D 4.511275e-02 mg C/g soil CORPSE
#> 48:    8        D 3.378580e-02 mg C/g soil CORPSE
#> 49:    9        D 2.659771e-02 mg C/g soil CORPSE
#> 50:   10        D 2.173648e-02 mg C/g soil CORPSE
#> 51:    1       EP 1.000000e-05 mg C/g soil CORPSE
#> 52:    2       EP 2.064157e-02 mg C/g soil CORPSE
#> 53:    3       EP 3.295457e-02 mg C/g soil CORPSE
#> 54:    4       EP 4.065573e-02 mg C/g soil CORPSE
#> 55:    5       EP 4.600692e-02 mg C/g soil CORPSE
#> 56:    6       EP 5.001089e-02 mg C/g soil CORPSE
#> 57:    7       EP 5.316460e-02 mg C/g soil CORPSE
#> 58:    8       EP 5.574091e-02 mg C/g soil CORPSE
#> 59:    9       EP 5.790272e-02 mg C/g soil CORPSE
#> 60:   10       EP 5.975404e-02 mg C/g soil CORPSE
#> 61:    1       EM 1.000000e-05 mg C/g soil CORPSE
#> 62:    2       EM 2.064157e-02 mg C/g soil CORPSE
#> 63:    3       EM 3.295457e-02 mg C/g soil CORPSE
#> 64:    4       EM 4.065573e-02 mg C/g soil CORPSE
#> 65:    5       EM 4.600692e-02 mg C/g soil CORPSE
#> 66:    6       EM 5.001089e-02 mg C/g soil CORPSE
#> 67:    7       EM 5.316460e-02 mg C/g soil CORPSE
#> 68:    8       EM 5.574091e-02 mg C/g soil CORPSE
#> 69:    9       EM 5.790272e-02 mg C/g soil CORPSE
#> 70:   10       EM 5.975404e-02 mg C/g soil CORPSE
#> 71:    1       IC 0.000000e+00 mg C/g soil CORPSE
#> 72:    2       IC 4.351559e+00 mg C/g soil CORPSE
#> 73:    3       IC 5.480883e+00 mg C/g soil CORPSE
#> 74:    4       IC 5.852951e+00 mg C/g soil CORPSE
#> 75:    5       IC 6.013613e+00 mg C/g soil CORPSE
#> 76:    6       IC 6.097708e+00 mg C/g soil CORPSE
#> 77:    7       IC 6.147916e+00 mg C/g soil CORPSE
#> 78:    8       IC 6.180759e+00 mg C/g soil CORPSE
#> 79:    9       IC 6.203701e+00 mg C/g soil CORPSE
#> 80:   10       IC 6.220538e+00 mg C/g soil CORPSE
#> 81:    1      Tot 1.810002e+01 mg C/g soil CORPSE
#> 82:    2      Tot 1.374846e+01 mg C/g soil CORPSE
#> 83:    3      Tot 1.261914e+01 mg C/g soil CORPSE
#> 84:    4      Tot 1.224707e+01 mg C/g soil CORPSE
#> 85:    5      Tot 1.208641e+01 mg C/g soil CORPSE
#> 86:    6      Tot 1.200231e+01 mg C/g soil CORPSE
#> 87:    7      Tot 1.195210e+01 mg C/g soil CORPSE
#> 88:    8      Tot 1.191926e+01 mg C/g soil CORPSE
#> 89:    9      Tot 1.189632e+01 mg C/g soil CORPSE
#> 90:   10      Tot 1.187948e+01 mg C/g soil CORPSE
#>     time variable        value       units   name