MEMC model configurations for MEND Wang et al. 2013

MEMC model configurations for MEND Wang et al. 2013

Usage

MEND_model

Format

an object created form configure_model.

name

MEND, the model name

table

table describing the user selected kinetics

params

the default MEMC parameter table

state

the default inital state values for MEMC

See also

Examples

MEND_model
#> $name
#> [1] "MEND"
#> 
#> $table
#>   model DOMdecomp POMdecomp MBdecay
#> 1  MEND        MM        MM      DD
#> 
#> $params
#>    parameter                                             description
#> 1        V.p         maximum specific decomposition rate for P by EP
#> 2        K.p         half-saturation constant for decomposition of P
#> 3        V.m         maximum specific decomposition rate for M by EM
#> 4        K.m         half-saturation constant for decomposition of M
#> 5        V.d       maximum specific uptake rate of D for growth of B
#> 6        K.d half-saturation constant of uptake of D for growth of B
#> 7        f.d                 fraction of decomposed P allocated to D
#> 8        g.d                       fraction of dead B allocated to D
#> 9       p.ep                     fraction of mR for production of EP
#> 10      p.em                     fraction of mR for production of EM
#> 11      r.ep                                     turnover rate of EP
#> 12      r.em                                     turnover rate of EM
#> 13     Q.max                           maximum DOC sorption capacity
#> 14     K.ads                                specific adsorption rate
#> 15     K.des                                         desorption rate
#> 16   dd.beta           strength of density dependent microbial decay
#> 17   Input.P                                              POM input 
#> 18   Input.D                                               DOM input
#> 19   Input.M                                                M input 
#> 20       CUE                                   carbon use efficiency
#>              units   value
#> 1  mgC mgC^-1 h^-1  14.000
#> 2     mgC / g soil  50.000
#> 3  mgC mgC^-1 h^-1   0.250
#> 4      mg C/g soil 250.000
#> 5  mgC mgC^-1 h^-1   3.000
#> 6      mg C/g soil   0.250
#> 7             <NA>   0.500
#> 8             <NA>   0.500
#> 9             <NA>   0.010
#> 10            <NA>   0.010
#> 11 mgC mgC^-1 h^-1   0.001
#> 12 mgC mgC^-1 h^-1   0.001
#> 13    mgC / g soil   3.400
#> 14 mgC mgC^-1 h^-1   0.006
#> 15 mgC mgC^-1 h^-1   0.001
#> 16            <NA>   2.000
#> 17            mg C   0.000
#> 18            mg C   0.000
#> 19            mg C   0.000
#> 20                   0.400
#> 
#> $state
#>        P        M        Q        B        D       EP       EM       IC 
#> 10.00000  5.00000  0.10000  2.00000  1.00000  0.00001  0.00001  0.00000 
#>      Tot 
#> 18.10002 
#> 
solve_model(mod = MEND_model, time = 1:10)
#>     time variable       value       units name
#>  1:    1        P 10.00000000 mg C/g soil MEND
#>  2:    2        P 10.95619098 mg C/g soil MEND
#>  3:    3        P 11.11889197 mg C/g soil MEND
#>  4:    4        P 11.14272636 mg C/g soil MEND
#>  5:    5        P 11.11848609 mg C/g soil MEND
#>  6:    6        P 11.07038243 mg C/g soil MEND
#>  7:    7        P 11.00786996 mg C/g soil MEND
#>  8:    8        P 10.93551315 mg C/g soil MEND
#>  9:    9        P 10.85583358 mg C/g soil MEND
#> 10:   10        P 10.77036028 mg C/g soil MEND
#> 11:    1        M  5.00000000 mg C/g soil MEND
#> 12:    2        M  5.01063569 mg C/g soil MEND
#> 13:    3        M  5.03438974 mg C/g soil MEND
#> 14:    4        M  5.06540822 mg C/g soil MEND
#> 15:    5        M  5.10137631 mg C/g soil MEND
#> 16:    6        M  5.14109633 mg C/g soil MEND
#> 17:    7        M  5.18385016 mg C/g soil MEND
#> 18:    8        M  5.22917041 mg C/g soil MEND
#> 19:    9        M  5.27673710 mg C/g soil MEND
#> 20:   10        M  5.32632365 mg C/g soil MEND
#> 21:    1        Q  0.10000000 mg C/g soil MEND
#> 22:    2        Q  0.10178079 mg C/g soil MEND
#> 23:    3        Q  0.10199580 mg C/g soil MEND
#> 24:    4        Q  0.10213966 mg C/g soil MEND
#> 25:    5        Q  0.10226937 mg C/g soil MEND
#> 26:    6        Q  0.10240270 mg C/g soil MEND
#> 27:    7        Q  0.10254738 mg C/g soil MEND
#> 28:    8        Q  0.10270683 mg C/g soil MEND
#> 29:    9        Q  0.10288220 mg C/g soil MEND
#> 30:   10        Q  0.10307326 mg C/g soil MEND
#> 31:    1        B  2.00000000 mg C/g soil MEND
#> 32:    2        B  0.78594185 mg C/g soil MEND
#> 33:    3        B  0.45507469 mg C/g soil MEND
#> 34:    4        B  0.32269533 mg C/g soil MEND
#> 35:    5        B  0.25374428 mg C/g soil MEND
#> 36:    6        B  0.21294150 mg C/g soil MEND
#> 37:    7        B  0.18709516 mg C/g soil MEND
#> 38:    8        B  0.17014106 mg C/g soil MEND
#> 39:    9        B  0.15888931 mg C/g soil MEND
#> 40:   10        B  0.15148798 mg C/g soil MEND
#> 41:    1        D  1.00000000 mg C/g soil MEND
#> 42:    2        D  0.05919225 mg C/g soil MEND
#> 43:    3        D  0.03309850 mg C/g soil MEND
#> 44:    4        D  0.02791116 mg C/g soil MEND
#> 45:    5        D  0.02745580 mg C/g soil MEND
#> 46:    6        D  0.02891942 mg C/g soil MEND
#> 47:    7        D  0.03125726 mg C/g soil MEND
#> 48:    8        D  0.03395304 mg C/g soil MEND
#> 49:    9        D  0.03670977 mg C/g soil MEND
#> 50:   10        D  0.03934787 mg C/g soil MEND
#> 51:    1       EP  0.00001000 mg C/g soil MEND
#> 52:    2       EP  0.01494581 mg C/g soil MEND
#> 53:    3       EP  0.02201630 mg C/g soil MEND
#> 54:    4       EP  0.02656158 mg C/g soil MEND
#> 55:    5       EP  0.02995278 mg C/g soil MEND
#> 56:    6       EP  0.03270196 mg C/g soil MEND
#> 57:    7       EP  0.03505710 mg C/g soil MEND
#> 58:    8       EP  0.03715743 mg C/g soil MEND
#> 59:    9       EP  0.03908886 mg C/g soil MEND
#> 60:   10       EP  0.04090793 mg C/g soil MEND
#> 61:    1       EM  0.00001000 mg C/g soil MEND
#> 62:    2       EM  0.01494581 mg C/g soil MEND
#> 63:    3       EM  0.02201630 mg C/g soil MEND
#> 64:    4       EM  0.02656158 mg C/g soil MEND
#> 65:    5       EM  0.02995278 mg C/g soil MEND
#> 66:    6       EM  0.03270196 mg C/g soil MEND
#> 67:    7       EM  0.03505710 mg C/g soil MEND
#> 68:    8       EM  0.03715743 mg C/g soil MEND
#> 69:    9       EM  0.03908886 mg C/g soil MEND
#> 70:   10       EM  0.04090793 mg C/g soil MEND
#> 71:    1       IC  0.00000000 mg C/g soil MEND
#> 72:    2       IC  1.15638683 mg C/g soil MEND
#> 73:    3       IC  1.31253671 mg C/g soil MEND
#> 74:    4       IC  1.38601611 mg C/g soil MEND
#> 75:    5       IC  1.43678258 mg C/g soil MEND
#> 76:    6       IC  1.47887370 mg C/g soil MEND
#> 77:    7       IC  1.51728588 mg C/g soil MEND
#> 78:    8       IC  1.55422064 mg C/g soil MEND
#> 79:    9       IC  1.59079031 mg C/g soil MEND
#> 80:   10       IC  1.62761110 mg C/g soil MEND
#> 81:    1      Tot 18.10002000 mg C/g soil MEND
#> 82:    2      Tot 16.94363317 mg C/g soil MEND
#> 83:    3      Tot 16.78748329 mg C/g soil MEND
#> 84:    4      Tot 16.71400389 mg C/g soil MEND
#> 85:    5      Tot 16.66323742 mg C/g soil MEND
#> 86:    6      Tot 16.62114630 mg C/g soil MEND
#> 87:    7      Tot 16.58273412 mg C/g soil MEND
#> 88:    8      Tot 16.54579936 mg C/g soil MEND
#> 89:    9      Tot 16.50922969 mg C/g soil MEND
#> 90:   10      Tot 16.47240890 mg C/g soil MEND
#>     time variable       value       units name