The default inital carbon pool values the basic MEMC model configurations from Wang et al. 2013

The default inital carbon pool values the basic MEMC model configurations from Wang et al. 2013

default_inital

Format

A vector of the intial states of the Carbon pools from MEND 2013 Wang et al. 2013.

'P'

'Particulate organic carbon or POC'

'M'

'Mineral associated organic carbon or MOC, note that this excludes Q'

'Q'

'Adsorbed phase of DOC'

'B'

'Microbial biomass carbon or MBC'

'D'

'Dissolved organic carbon or DOC'

'EP'

'Enzyme for decomposition of P'

'EM'

'Ensyme for decomposition of M'

'IC'

'Accumulated CO2 flux'

'Tot'

'Total carbon in the patch'

References

Wang et al. 2013

Examples

default_inital
#>        P        M        Q        B        D       EP       EM       IC 
#> 10.00000  5.00000  0.10000  2.00000  1.00000  0.00001  0.00001  0.00000 
#>      Tot 
#> 18.10002 
print(default_inital)
#>        P        M        Q        B        D       EP       EM       IC 
#> 10.00000  5.00000  0.10000  2.00000  1.00000  0.00001  0.00001  0.00000 
#>      Tot 
#> 18.10002