The parameter values for the basic MEMC model configurations from Wang et al. 2013

The parameter values for the basic MEMC model configurations from Wang et al. 2013

data(default_params)

Format

A data table consisting of 4 columns with inputs from Wang et al. 2013.

parameter

String character of the default MEND parameters.

description

String character describing the parameter.

units

String character of the parameter units.

value

Numeric values taken from the table 2 of Wang et al. 2013

References

Wang et al. 2013

See also

Examples

default_params
#>    parameter                                             description
#> 1        V.p         maximum specific decomposition rate for P by EP
#> 2        K.p         half-saturation constant for decomposition of P
#> 3        V.m         maximum specific decomposition rate for M by EM
#> 4        K.m         half-saturation constant for decomposition of M
#> 5        V.d       maximum specific uptake rate of D for growth of B
#> 6        K.d half-saturation constant of uptake of D for growth of B
#> 7        m.r                     specific maintenance factor or rate
#> 8        E.c                                   carbon use efficiency
#> 9        f.d                 fraction of decomposed P allocated to D
#> 10       g.d                       fraction of dead B allocated to D
#> 11      p.ep                     fraction of mR for production of EP
#> 12      p.em                     fraction of mR for production of EM
#> 13      r.ep                                     turnover rate of EP
#> 14      r.em                                     turnover rate of EM
#> 15     Q.max                           maximum DOC sorption capacity
#> 16     K.ads                                specific adsorption rate
#> 17     K.des                                         desorption rate
#> 18      K.ba                                        binding affinity
#> 19       I.p                                         input rate of P
#> 20       I.d                                         input rate of D
#> 21      fI.d                                       ratio of ID to IP
#>              units   value
#> 1  mgC mgC^-1 h^-1 2.5e+00
#> 2     mgC / g soil 5.0e+01
#> 3  mgC mgC^-1 h^-1 1.0e+00
#> 4      mg C/g soil 2.5e+02
#> 5  mgC mgC^-1 h^-1 5.0e-04
#> 6      mg C/g soil 2.6e-01
#> 7  mgC mgC^-1 h^-1 2.8e-04
#> 8             <NA> 4.7e-01
#> 9             <NA> 5.0e-01
#> 10            <NA> 5.0e-01
#> 11            <NA> 1.0e-02
#> 12            <NA> 1.0e-02
#> 13 mgC mgC^-1 h^-1 1.0e-03
#> 14 mgC mgC^-1 h^-1 1.0e-03
#> 15    mgC / g soil 1.7e+00
#> 16 mgC mgC^-1 h^-1 6.0e-03
#> 17 mgC mgC^-1 h^-1 1.0e-03
#> 18   (mgC/soil)^-1 6.0e+00
#> 19 mgC mgC^-1 h^-1 8.0e-05
#> 20 mgC mgC^-1 h^-1 8.0e-05
#> 21            <NA> 1.0e+00