Solve a MEMC configuration

Solve a MEMC configuration

Usage

memc_solve(mod, time, params = NULL, state = NULL, ...)

Arguments

mod

model object, for example one of the list entries returned by memc_all_models

time

daily time steps for model run

params

default set to NULL, will then use the parameter table read in with the mod object

state

default set to NULL, will then use the state read read in with the mod object

...

additional arguments passed to ode

Value

A long-formatted data.table of the simulation results, with columns:

time

Time point, days

variable

Name of model carbon pool

value

Value of the pool

units

Units of the pool value

name

Model name used

See also

Examples

memc_solve(MEND_model, time = 0:10)
#>      time variable        value       units   name
#>     <num>   <fctr>        <num>      <char> <char>
#>  1:     0      POM 1.000000e+01 mg C/g soil   MEND
#>  2:     1      POM 1.085536e+01 mg C/g soil   MEND
#>  3:     2      POM 1.115375e+01 mg C/g soil   MEND
#>  4:     3      POM 1.122363e+01 mg C/g soil   MEND
#>  5:     4      POM 1.120608e+01 mg C/g soil   MEND
#>  6:     5      POM 1.115458e+01 mg C/g soil   MEND
#>  7:     6      POM 1.108957e+01 mg C/g soil   MEND
#>  8:     7      POM 1.101908e+01 mg C/g soil   MEND
#>  9:     8      POM 1.094641e+01 mg C/g soil   MEND
#> 10:     9      POM 1.087294e+01 mg C/g soil   MEND
#> 11:    10      POM 1.079933e+01 mg C/g soil   MEND
#> 12:     0      MOM 5.000000e+00 mg C/g soil   MEND
#> 13:     1      MOM 5.012398e+00 mg C/g soil   MEND
#> 14:     2      MOM 5.040316e+00 mg C/g soil   MEND
#> 15:     3      MOM 5.074505e+00 mg C/g soil   MEND
#> 16:     4      MOM 5.111070e+00 mg C/g soil   MEND
#> 17:     5      MOM 5.148468e+00 mg C/g soil   MEND
#> 18:     6      MOM 5.186081e+00 mg C/g soil   MEND
#> 19:     7      MOM 5.223655e+00 mg C/g soil   MEND
#> 20:     8      MOM 5.261082e+00 mg C/g soil   MEND
#> 21:     9      MOM 5.298312e+00 mg C/g soil   MEND
#> 22:    10      MOM 5.335321e+00 mg C/g soil   MEND
#> 23:     0      QOM 1.000000e-01 mg C/g soil   MEND
#> 24:     1      QOM 1.012582e-01 mg C/g soil   MEND
#> 25:     2      QOM 1.016135e-01 mg C/g soil   MEND
#> 26:     3      QOM 1.020186e-01 mg C/g soil   MEND
#> 27:     4      QOM 1.025091e-01 mg C/g soil   MEND
#> 28:     5      QOM 1.031229e-01 mg C/g soil   MEND
#> 29:     6      QOM 1.038915e-01 mg C/g soil   MEND
#> 30:     7      QOM 1.048370e-01 mg C/g soil   MEND
#> 31:     8      QOM 1.059725e-01 mg C/g soil   MEND
#> 32:     9      QOM 1.073049e-01 mg C/g soil   MEND
#> 33:    10      QOM 1.088369e-01 mg C/g soil   MEND
#> 34:     0       MB 2.000000e+00 mg C/g soil   MEND
#> 35:     1       MB 9.347011e-01 mg C/g soil   MEND
#> 36:     2       MB 3.618926e-01 mg C/g soil   MEND
#> 37:     3       MB 1.436626e-01 mg C/g soil   MEND
#> 38:     4       MB 5.935201e-02 mg C/g soil   MEND
#> 39:     5       MB 2.582606e-02 mg C/g soil   MEND
#> 40:     6       MB 1.190014e-02 mg C/g soil   MEND
#> 41:     7       MB 5.802372e-03 mg C/g soil   MEND
#> 42:     8       MB 2.981106e-03 mg C/g soil   MEND
#> 43:     9       MB 1.604872e-03 mg C/g soil   MEND
#> 44:    10       MB 9.003539e-04 mg C/g soil   MEND
#> 45:     0      DOM 1.000000e+00 mg C/g soil   MEND
#> 46:     1      DOM 6.365408e-02 mg C/g soil   MEND
#> 47:     2      DOM 6.954756e-02 mg C/g soil   MEND
#> 48:     3      DOM 8.099383e-02 mg C/g soil   MEND
#> 49:     4      DOM 9.902432e-02 mg C/g soil   MEND
#> 50:     5      DOM 1.232404e-01 mg C/g soil   MEND
#> 51:     6      DOM 1.521146e-01 mg C/g soil   MEND
#> 52:     7      DOM 1.839678e-01 mg C/g soil   MEND
#> 53:     8      DOM 2.175026e-01 mg C/g soil   MEND
#> 54:     9      DOM 2.518673e-01 mg C/g soil   MEND
#> 55:    10      DOM 2.865449e-01 mg C/g soil   MEND
#> 56:     0       EP 1.000000e-05 mg C/g soil   MEND
#> 57:     1       EP 1.796429e-02 mg C/g soil   MEND
#> 58:     2       EP 2.517575e-02 mg C/g soil   MEND
#> 59:     3       EP 2.797598e-02 mg C/g soil   MEND
#> 60:     4       EP 2.908749e-02 mg C/g soil   MEND
#> 61:     5       EP 2.953941e-02 mg C/g soil   MEND
#> 62:     6       EP 2.972442e-02 mg C/g soil   MEND
#> 63:     7       EP 2.979606e-02 mg C/g soil   MEND
#> 64:     8       EP 2.981688e-02 mg C/g soil   MEND
#> 65:     9       EP 2.981363e-02 mg C/g soil   MEND
#> 66:    10       EP 2.979841e-02 mg C/g soil   MEND
#> 67:     0       EM 1.000000e-05 mg C/g soil   MEND
#> 68:     1       EM 1.796429e-02 mg C/g soil   MEND
#> 69:     2       EM 2.517575e-02 mg C/g soil   MEND
#> 70:     3       EM 2.797598e-02 mg C/g soil   MEND
#> 71:     4       EM 2.908749e-02 mg C/g soil   MEND
#> 72:     5       EM 2.953941e-02 mg C/g soil   MEND
#> 73:     6       EM 2.972442e-02 mg C/g soil   MEND
#> 74:     7       EM 2.979606e-02 mg C/g soil   MEND
#> 75:     8       EM 2.981688e-02 mg C/g soil   MEND
#> 76:     9       EM 2.981363e-02 mg C/g soil   MEND
#> 77:    10       EM 2.979841e-02 mg C/g soil   MEND
#> 78:     0       IC 0.000000e+00 mg C/g soil   MEND
#> 79:     1       IC 1.096718e+00 mg C/g soil   MEND
#> 80:     2       IC 1.322545e+00 mg C/g soil   MEND
#> 81:     3       IC 1.419253e+00 mg C/g soil   MEND
#> 82:     4       IC 1.463805e+00 mg C/g soil   MEND
#> 83:     5       IC 1.485707e+00 mg C/g soil   MEND
#> 84:     6       IC 1.497016e+00 mg C/g soil   MEND
#> 85:     7       IC 1.503081e+00 mg C/g soil   MEND
#> 86:     8       IC 1.506442e+00 mg C/g soil   MEND
#> 87:     9       IC 1.508364e+00 mg C/g soil   MEND
#> 88:    10       IC 1.509494e+00 mg C/g soil   MEND
#> 89:     0      Tot 1.810002e+01 mg C/g soil   MEND
#> 90:     1      Tot 1.700330e+01 mg C/g soil   MEND
#> 91:     2      Tot 1.677747e+01 mg C/g soil   MEND
#> 92:     3      Tot 1.668077e+01 mg C/g soil   MEND
#> 93:     4      Tot 1.663621e+01 mg C/g soil   MEND
#> 94:     5      Tot 1.661431e+01 mg C/g soil   MEND
#> 95:     6      Tot 1.660300e+01 mg C/g soil   MEND
#> 96:     7      Tot 1.659694e+01 mg C/g soil   MEND
#> 97:     8      Tot 1.659358e+01 mg C/g soil   MEND
#> 98:     9      Tot 1.659166e+01 mg C/g soil   MEND
#> 99:    10      Tot 1.659053e+01 mg C/g soil   MEND
#>      time variable        value       units   name