MEMC model configurations for MEND Wang et al. 2013

MEMC model configurations for MEND Wang et al. 2013

Usage

COMISSION_model

Format

an object created form configure_model.

name

COMISSION, the model name

table

table describing the user selected kinetics

params

the default MEMC parameter table

state

the default inital state values for MEMC

See also

Examples

COMISSION_model
#> $name
#> [1] "COMISSION"
#> 
#> $table
#>       model DOMdecomp POMdecomp MBdecay
#> 1 COMISSION        MM       RMM      DD
#> 
#> $params
#>    parameter                                             description
#> 1        V.p         maximum specific decomposition rate for P by EP
#> 2        K.p         half-saturation constant for decomposition of P
#> 3        V.m         maximum specific decomposition rate for M by EM
#> 4        K.m         half-saturation constant for decomposition of M
#> 5        V.d       maximum specific uptake rate of D for growth of B
#> 6        K.d half-saturation constant of uptake of D for growth of B
#> 7        f.d                 fraction of decomposed P allocated to D
#> 8        g.d                       fraction of dead B allocated to D
#> 9       p.ep                     fraction of mR for production of EP
#> 10      p.em                     fraction of mR for production of EM
#> 11      r.ep                                     turnover rate of EP
#> 12      r.em                                     turnover rate of EM
#> 13     Q.max                           maximum DOC sorption capacity
#> 14     K.ads                                specific adsorption rate
#> 15     K.des                                         desorption rate
#> 16   dd.beta           strength of density dependent microbial decay
#> 17   Input.P                                              POM input 
#> 18   Input.D                                               DOM input
#> 19   Input.M                                                M input 
#> 20       CUE                                   carbon use efficiency
#>              units   value
#> 1  mgC mgC^-1 h^-1  14.000
#> 2     mgC / g soil  50.000
#> 3  mgC mgC^-1 h^-1   0.250
#> 4      mg C/g soil 250.000
#> 5  mgC mgC^-1 h^-1   3.000
#> 6      mg C/g soil   0.250
#> 7             <NA>   0.500
#> 8             <NA>   0.500
#> 9             <NA>   0.010
#> 10            <NA>   0.010
#> 11 mgC mgC^-1 h^-1   0.001
#> 12 mgC mgC^-1 h^-1   0.001
#> 13    mgC / g soil   3.400
#> 14 mgC mgC^-1 h^-1   0.006
#> 15 mgC mgC^-1 h^-1   0.001
#> 16            <NA>   2.000
#> 17            mg C   0.000
#> 18            mg C   0.000
#> 19            mg C   0.000
#> 20                   0.400
#> 
#> $state
#>        P        M        Q        B        D       EP       EM       IC 
#> 10.00000  5.00000  0.10000  2.00000  1.00000  0.00001  0.00001  0.00000 
#>      Tot 
#> 18.10002 
#> 
solve_model(mod = COMISSION_model, time = 1:10)
#>     time variable       value       units      name
#>  1:    1        P 10.00000000 mg C/g soil COMISSION
#>  2:    2        P 10.95175351 mg C/g soil COMISSION
#>  3:    3        P 11.10452319 mg C/g soil COMISSION
#>  4:    4        P 11.11540904 mg C/g soil COMISSION
#>  5:    5        P 11.07624826 mg C/g soil COMISSION
#>  6:    6        P 11.01180379 mg C/g soil COMISSION
#>  7:    7        P 10.93186649 mg C/g soil COMISSION
#>  8:    8        P 10.84122359 mg C/g soil COMISSION
#>  9:    9        P 10.74255233 mg C/g soil COMISSION
#> 10:   10        P 10.63749510 mg C/g soil COMISSION
#> 11:    1        M  5.00000000 mg C/g soil COMISSION
#> 12:    2        M  5.01292828 mg C/g soil COMISSION
#> 13:    3        M  5.04193185 mg C/g soil COMISSION
#> 14:    4        M  5.07983004 mg C/g soil COMISSION
#> 15:    5        M  5.12375290 mg C/g soil COMISSION
#> 16:    6        M  5.17221290 mg C/g soil COMISSION
#> 17:    7        M  5.22431859 mg C/g soil COMISSION
#> 18:    8        M  5.27949043 mg C/g soil COMISSION
#> 19:    9        M  5.33733179 mg C/g soil COMISSION
#> 20:   10        M  5.39756102 mg C/g soil COMISSION
#> 21:    1        Q  0.10000000 mg C/g soil COMISSION
#> 22:    2        Q  0.10178269 mg C/g soil COMISSION
#> 23:    3        Q  0.10200319 mg C/g soil COMISSION
#> 24:    4        Q  0.10215718 mg C/g soil COMISSION
#> 25:    5        Q  0.10230202 mg C/g soil COMISSION
#> 26:    6        Q  0.10245543 mg C/g soil COMISSION
#> 27:    7        Q  0.10262471 mg C/g soil COMISSION
#> 28:    8        Q  0.10281263 mg C/g soil COMISSION
#> 29:    9        Q  0.10301948 mg C/g soil COMISSION
#> 30:   10        Q  0.10324417 mg C/g soil COMISSION
#> 31:    1        B  2.00000000 mg C/g soil COMISSION
#> 32:    2        B  0.78637696 mg C/g soil COMISSION
#> 33:    3        B  0.45639096 mg C/g soil COMISSION
#> 34:    4        B  0.32507878 mg C/g soil COMISSION
#> 35:    5        B  0.25730143 mg C/g soil COMISSION
#> 36:    6        B  0.21773136 mg C/g soil COMISSION
#> 37:    7        B  0.19314393 mg C/g soil COMISSION
#> 38:    8        B  0.17744680 mg C/g soil COMISSION
#> 39:    9        B  0.16742303 mg C/g soil COMISSION
#> 40:   10        B  0.16119403 mg C/g soil COMISSION
#> 41:    1        D  1.00000000 mg C/g soil COMISSION
#> 42:    2        D  0.05979623 mg C/g soil COMISSION
#> 43:    3        D  0.03442059 mg C/g soil COMISSION
#> 44:    4        D  0.03007891 mg C/g soil COMISSION
#> 45:    5        D  0.03049065 mg C/g soil COMISSION
#> 46:    6        D  0.03277811 mg C/g soil COMISSION
#> 47:    7        D  0.03584400 mg C/g soil COMISSION
#> 48:    8        D  0.03913554 mg C/g soil COMISSION
#> 49:    9        D  0.04233826 mg C/g soil COMISSION
#> 50:   10        D  0.04527398 mg C/g soil COMISSION
#> 51:    1       EP  0.00001000 mg C/g soil COMISSION
#> 52:    2       EP  0.01494741 mg C/g soil COMISSION
#> 53:    3       EP  0.02202817 mg C/g soil COMISSION
#> 54:    4       EP  0.02659549 mg C/g soil COMISSION
#> 55:    5       EP  0.03002221 mg C/g soil COMISSION
#> 56:    6       EP  0.03282133 mg C/g soil COMISSION
#> 57:    7       EP  0.03524134 mg C/g soil COMISSION
#> 58:    8       EP  0.03742159 mg C/g soil COMISSION
#> 59:    9       EP  0.03944779 mg C/g soil COMISSION
#> 60:   10       EP  0.04137595 mg C/g soil COMISSION
#> 61:    1       EM  0.00001000 mg C/g soil COMISSION
#> 62:    2       EM  0.01494741 mg C/g soil COMISSION
#> 63:    3       EM  0.02202817 mg C/g soil COMISSION
#> 64:    4       EM  0.02659549 mg C/g soil COMISSION
#> 65:    5       EM  0.03002221 mg C/g soil COMISSION
#> 66:    6       EM  0.03282133 mg C/g soil COMISSION
#> 67:    7       EM  0.03524134 mg C/g soil COMISSION
#> 68:    8       EM  0.03742159 mg C/g soil COMISSION
#> 69:    9       EM  0.03944779 mg C/g soil COMISSION
#> 70:   10       EM  0.04137595 mg C/g soil COMISSION
#> 71:    1       IC  0.00000000 mg C/g soil COMISSION
#> 72:    2       IC  1.15748751 mg C/g soil COMISSION
#> 73:    3       IC  1.31669388 mg C/g soil COMISSION
#> 74:    4       IC  1.39427506 mg C/g soil COMISSION
#> 75:    5       IC  1.44988034 mg C/g soil COMISSION
#> 76:    6       IC  1.49739575 mg C/g soil COMISSION
#> 77:    7       IC  1.54173960 mg C/g soil COMISSION
#> 78:    8       IC  1.58506783 mg C/g soil COMISSION
#> 79:    9       IC  1.62845953 mg C/g soil COMISSION
#> 80:   10       IC  1.67249979 mg C/g soil COMISSION
#> 81:    1      Tot 18.10002000 mg C/g soil COMISSION
#> 82:    2      Tot 16.94253249 mg C/g soil COMISSION
#> 83:    3      Tot 16.78332612 mg C/g soil COMISSION
#> 84:    4      Tot 16.70574494 mg C/g soil COMISSION
#> 85:    5      Tot 16.65013966 mg C/g soil COMISSION
#> 86:    6      Tot 16.60262425 mg C/g soil COMISSION
#> 87:    7      Tot 16.55828040 mg C/g soil COMISSION
#> 88:    8      Tot 16.51495217 mg C/g soil COMISSION
#> 89:    9      Tot 16.47156047 mg C/g soil COMISSION
#> 90:   10      Tot 16.42752021 mg C/g soil COMISSION
#>     time variable       value       units      name